requirements/runtime.txt
matplotlib
mmcv-full==1.3.0
mmpycocotools
numpy
six
terminaltables
torch>=1.7.1,<=1.8.1
torchvision>=0.8.1,<=0.9.1
pytorchcv==0.0.55
tqdm
onnx>=1.7.0
onnxoptimizer
onnxruntime>=1.1.2
lxml
Polygon3==3.0.8
editdistance
pyclipper
